Add files via upload
In this project, I have tried to elaborate on visualization techniques frequently practiced in many industries, academia, and government corporations. The sheer breadth and depth of the *ggplot2* and its counterparts are beyond the scope of this project. But, soon this project will be updated with more sophisticated and interactive visualization techniques. R programming language has rich and powerful data visualization capabilities. While tools like Excel, Power BI, and Tableau are often the go-to solutions for data visualizations, none of these tools can compete with R in terms of the sheer breadth of, and control over, crafted data visualizations. We will mostly be focusing on the visualization tools provided by the mighty *ggplot2* library in R. The *ggplot2* library adopts the *grammar of graphics* which essentially means, it is possible to create hundreds of plots by simply leveraging the handful of verbs, nouns, and adjectives of graphical representation.